requirements.txt
Flask==2.0.2
mysql_connector_repackaged==0.3.1
numpy==1.17.4
pandas==1.3.2
gunicorn==20.1.0